refactor(dashboard): split OrganizationManagement god component (#80, part 2)
Resolves smell GM-8. The page was 904 LOC owning org CRUD, members CRUD,
invitations CRUD, and an embedded `AuditLogs` component starting at
line ~855.

Split:
- `OrganizationManagement.tsx` 904 → **498 LOC** (orchestrator only)
- `components/AuditLogs.tsx` (NEW, 61 LOC) — the embedded component
  is now in its own file (acceptance criterion satisfied)
- `components/OrganizationEditPanel.tsx` (NEW, 54 LOC) — create-org
  form with controlled state
- `components/OrganizationMembersPanel.tsx` (NEW, 236 LOC) — members
  table + invite form + delegates to InvitationsPanel
- `components/OrganizationInvitationsPanel.tsx` (NEW, 136 LOC) —
  invitations table + filter tabs + audit modal (auditOpen state
  stays inside this panel since the modal is self-contained)
- `components/ConfirmationDialog.tsx` (NEW, 56 LOC) — extracted shared
  reusable dialog (used twice: mode-switch + delete-org confirmations)
  to fit orchestrator under the 500 LOC budget

All sub-files ≤ 500 LOC. Default export of `OrganizationManagement`
preserved → `UserAccountButton.tsx` import unchanged.

Design call: passed `selectedOrg`, `members`, `invitations`, and all
fetch callbacks down via explicit props rather than introducing a new
context — keeps data flow visible in a leaf-page component split.

Hardest extraction: `auditOpen` state was referenced from inside
invitation rows (the "View Audit" button). Since it controls a modal
that also needs `selectedOrg.id`, it naturally belongs entirely
inside `OrganizationInvitationsPanel` — no cross-panel shared state
needed.

Verification:
- bun run test: **294/294 passed** (49 suites; the dedicated
  `OrganizationManagement.test.tsx` 26-`render(...)` suite passes
  unmodified — strongest regression net for this split)
- bun run build: clean, 2.30s

Closes #80 acceptance: no file under components/ for these two pages
exceeds 500 LOC; AuditLogs lives in its own file; existing
functionality verified by the unmodified test suite.
